Hotfix * Fixed Doors and Keys challenge. * Fixed inability to use user bot script outside of game folder in multiplayer. * Separated game files.
