+++ game/challenges/doorsandkeys.json
@@ -15,13 +15,13 @@
     {
       "map": 1,
       "steps": 200,
-      "start": [1, 23],
+      "start": [1, 1],
       "gold": 15
     },
     {
       "map": 2,
       "steps": 400,
-      "start": [1, 1],
+      "start": [4, 3],
       "gold": 28
     }
   ],
@@ -52,12 +52,12 @@
       "####                    ####",
       "####                    ####",
       "#######   ########   #######",
-      "#############11###DDD#######",
+      "#######DDD###11###   #######",
       "#######   ###11###   #######",
-      "####                    ####",
-      "####                    ####",
-      "#1                        1#",
-      "#                          #",
+      "####           #        ####",
+      "####           #        ####",
+      "#1             #          1#",
+      "#              #           #",
       "#           ####           #",
       "#           ####           #",
       "#           ####           #",
@@ -75,12 +75,12 @@
       "#                          #",
       "#        ##########        #",
       "#        #11111111#        #",
-      "#  #DD#  #        #1 #DD# 1#",
+      "#  #  #  #        #1 #  # 1#",
       "####  ####        ####  ####",
       "   #  ####        ####  #   ",
-      "   #                    #   ",
-      "   #                    #   ",
-      "   #DD#              #DD#   ",
+      "   #  D              #  #   ",
+      "   #  D              #  #   ",
+      "   #  #              #  #   ",
       "   #  ####        ####  #   ",
       "####  ####        ####  ####",
       "#1      1#        #1      1#",
